wall

+++ srcs/isitwall/dis_wall.c
@@ -6,7 +6,7 @@
 /*   By: pganglof <pganglof@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/27 17:47:21 by pganglof          #+#    #+#             */
-/*   Updated: 2019/11/28 15:28:18 by pganglof         ###   ########.fr       */
+/*   Updated: 2019/11/28 18:43:14 by pganglof         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,13 +38,13 @@ double	dis_wall(t_pos *gamer, t_pos *wall)
 
 	printf("pow1 :%G\n", pow(*(int*)gamer->unitx - *(float*)wall->unitx, 2));
 	
-	printf("dis : %G\n", sqrt((double)(pow(*(int*)gamer->unitx - *(float*)wall->unitx, 2)
-	+ pow(*(int*)gamer->unity - *(float*)wall->unity, 2)))
-	* cos(gamer->beta));
+	printf("dis : %G\n", sqrt(pow((double)(*(int*)gamer->unitx - *(float*)wall->unitx), 2)
+	+ pow((double)(*(int*)gamer->unity - *(float*)wall->unity), 2))
+	* cos(gamer->beta * M_PI / 180));
 
-	return (sqrt(pow(*(int*)gamer->unitx - *(float*)wall->unitx, 2)
-	+ pow(*(int*)gamer->unity - *(float*)wall->unity, 2))
-	* cos(gamer->beta));
+	return (sqrt(pow((double)(*(int*)gamer->unitx - *(float*)wall->unitx), 2)
+	+ pow((double)(*(int*)gamer->unity - *(float*)wall->unity), 2))
+	* cos(gamer->beta * M_PI / 180));
 
 /*	return ((double)sqrt((double)((double)(ft_iterative_power(*(int*)gamer->unitx - *(float*)wall->unitx, 2)
 	+ (double)(ft_iterative_power(*(int*)gamer->unity - *(float*)wall->unity, 2))))
